# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DG Cluster Assortment Advisor API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/skus": {
      "get": {
        "tags": [
          "SKUs"
        ],
        "summary": "Get Skus",
        "description": "Retrieve Snacks SKU list with performance metrics and color-coded status badges.",
        "operationId": "get_skus_api_v1_skus_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search by SKU name, code, or subcategory",
              "title": "Search"
            },
            "description": "Search by SKU name, code, or subcategory"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Category filter (e.g. Snacks)",
              "title": "Category"
            },
            "description": "Category filter (e.g. Snacks)"
          },
          {
            "name": "subcategory",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by subcategory",
              "title": "Subcategory"
            },
            "description": "Filter by subcategory"
          },
          {
            "name": "status_badge",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status badge (GROW, MAINTAIN, SWAP, REDUCE)",
              "title": "Status Badge"
            },
            "description": "Filter by status badge (GROW, MAINTAIN, SWAP, REDUCE)"
          },
          {
            "name": "brand_tier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by brand tier (Private Brand, National Brand)",
              "title": "Brand Tier"
            },
            "description": "Filter by brand tier (Private Brand, National Brand)"
          },
          {
            "name": "cluster_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by cluster code (e.g. STV-CLUSTER)",
              "title": "Cluster Code"
            },
            "description": "Filter by cluster code (e.g. STV-CLUSTER)"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkuListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "SKUs"
        ],
        "summary": "Create Sku",
        "description": "Create a new Snacks SKU.",
        "operationId": "create_sku_api_v1_skus_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkuCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkuResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/skus/{sku_id}": {
      "get": {
        "tags": [
          "SKUs"
        ],
        "summary": "Get Sku",
        "description": "Retrieve a single SKU by ID or SKU Code.",
        "operationId": "get_sku_api_v1_skus__sku_id__get",
        "parameters": [
          {
            "name": "sku_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sku Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkuResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "SKUs"
        ],
        "summary": "Update Sku",
        "description": "Update SKU attributes or status badge.",
        "operationId": "update_sku_api_v1_skus__sku_id__patch",
        "parameters": [
          {
            "name": "sku_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sku Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkuUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkuResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/assortment-plans/submit": {
      "post": {
        "tags": [
          "Assortment Plans"
        ],
        "summary": "Submit Plan",
        "description": "Commit and approve selected assortment plan with audit log.",
        "operationId": "submit_plan_api_v1_assortment_plans_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssortmentPlanSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssortmentPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assortment-plans/{audit_id}": {
      "get": {
        "tags": [
          "Assortment Plans"
        ],
        "summary": "Get Plan",
        "description": "Retrieve historical audit trail record by audit ID or plan ID.",
        "operationId": "get_plan_api_v1_assortment_plans__audit_id__get",
        "parameters": [
          {
            "name": "audit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Audit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssortmentPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssortmentPlanResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "audit_id": {
            "type": "string",
            "title": "Audit Id"
          },
          "cluster_id": {
            "type": "string",
            "title": "Cluster Id"
          },
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "submitted_by": {
            "type": "string",
            "title": "Submitted By"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "total_sku_actions": {
            "type": "integer",
            "title": "Total Sku Actions"
          },
          "guardrail_status": {
            "type": "string",
            "title": "Guardrail Status"
          },
          "summary_snapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary Snapshot"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "audit_trail_summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanAuditSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "audit_id",
          "cluster_id",
          "scenario_id",
          "submitted_by",
          "status",
          "total_sku_actions",
          "guardrail_status",
          "timestamp",
          "created_at",
          "updated_at"
        ],
        "title": "AssortmentPlanResponse"
      },
      "AssortmentPlanSubmitRequest": {
        "properties": {
          "scenario_code": {
            "type": "string",
            "title": "Scenario Code"
          },
          "cluster_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Code",
            "default": "STV-CLUSTER"
          },
          "submitted_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted By",
            "default": "Category Manager"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "scenario_code"
        ],
        "title": "AssortmentPlanSubmitRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanAuditSummary": {
        "properties": {
          "audit_id": {
            "type": "string",
            "title": "Audit Id"
          },
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "scenario_code": {
            "type": "string",
            "title": "Scenario Code"
          },
          "scenario_name": {
            "type": "string",
            "title": "Scenario Name"
          },
          "cluster_code": {
            "type": "string",
            "title": "Cluster Code"
          },
          "submitted_by": {
            "type": "string",
            "title": "Submitted By"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "guardrail_status": {
            "type": "string",
            "title": "Guardrail Status"
          },
          "total_sku_actions": {
            "type": "integer",
            "title": "Total Sku Actions"
          },
          "summary_snapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary Snapshot"
          },
          "action_breakdown": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Breakdown"
          },
          "actions": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SkuActionDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actions"
          }
        },
        "type": "object",
        "required": [
          "audit_id",
          "plan_id",
          "scenario_code",
          "scenario_name",
          "cluster_code",
          "submitted_by",
          "timestamp",
          "guardrail_status",
          "total_sku_actions"
        ],
        "title": "PlanAuditSummary"
      },
...
      "SkuActionDetail": {
        "properties": {
          "sku_id": {
            "type": "string",
            "title": "Sku Id"
          },
          "sku_code": {
            "type": "string",
            "title": "Sku Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "action_type": {
            "type": "string",
            "title": "Action Type"
          }
        },
        "type": "object",
        "required": [
          "sku_id",
          "sku_code",
          "name",
          "action_type"
        ],
        "title": "SkuActionDetail"
      },
      "SkuCreate": {
        "properties": {
          "sku_code": {
            "type": "string",
            "title": "Sku Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "subcategory": {
            "type": "string",
            "title": "Subcategory"
          },
          "brand_tier": {
            "type": "string",
            "title": "Brand Tier"
          },
          "sales_per_lin_ft": {
            "type": "number",
            "title": "Sales Per Lin Ft"
          },
          "margin_pct": {
            "type": "number",
            "title": "Margin Pct"
          },
          "weekly_unit_velocity": {
            "type": "number",
            "title": "Weekly Unit Velocity"
          },
          "in_stock_pct": {
            "type": "number",
            "title": "In Stock Pct"
          },
          "shelf_linear_ft": {
            "type": "number",
            "title": "Shelf Linear Ft"
          },
          "status_badge": {
            "type": "string",
            "title": "Status Badge"
          },
          "cluster_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Id"
          }
        },
        "type": "object",
        "required": [
          "sku_code",
          "name",
          "subcategory",
          "brand_tier",
          "sales_per_lin_ft",
          "margin_pct",
          "weekly_unit_velocity",
          "in_stock_pct",
          "shelf_linear_ft",
          "status_badge"
        ],
        "title": "SkuCreate"
      },
      "SkuListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SkuResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "SkuListResponse"
      },
      "SkuResponse": {
        "properties": {
          "sku_code": {
            "type": "string",
            "title": "Sku Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "subcategory": {
            "type": "string",
            "title": "Subcategory"
          },
          "brand_tier": {
            "type": "string",
            "title": "Brand Tier"
          },
          "sales_per_lin_ft": {
            "type": "number",
            "title": "Sales Per Lin Ft"
          },
          "margin_pct": {
            "type": "number",
            "title": "Margin Pct"
          },
          "weekly_unit_velocity": {
            "type": "number",
            "title": "Weekly Unit Velocity"
          },
          "in_stock_pct": {
            "type": "number",
            "title": "In Stock Pct"
          },
          "shelf_linear_ft": {
            "type": "number",
            "title": "Shelf Linear Ft"
          },
          "status_badge": {
            "type": "string",
            "title": "Status Badge"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "cluster_id": {
            "type": "string",
            "title": "Cluster Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "sku_code",
          "name",
          "subcategory",
          "brand_tier",
          "sales_per_lin_ft",
          "margin_pct",
          "weekly_unit_velocity",
          "in_stock_pct",
          "shelf_linear_ft",
          "status_badge",
          "id",
          "cluster_id"
        ],
        "title": "SkuResponse"
      },
      "SkuUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "subcategory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subcategory"
          },
          "brand_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Tier"
          },
          "sales_per_lin_ft": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sales Per Lin Ft"
          },
          "margin_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Margin Pct"
          },
          "weekly_unit_velocity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Unit Velocity"
          },
          "in_stock_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Stock Pct"
          },
          "shelf_linear_ft": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shelf Linear Ft"
          },
          "status_badge": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Badge"
          }
        },
        "type": "object",
        "title": "SkuUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}